Move option processing into vine_worker_options.[ch]
